+++ crates/key-server/src/tests/e2e.rs
@@ -10,8 +10,8 @@ use crate::sui_rpc_client::SuiRpcClient;
 use crate::tests::externals::get_key;
 use crate::tests::whitelist::{add_user_to_whitelist, create_whitelist, whitelist_create_ptb};
 use crate::tests::SealTestCluster;
+use crate::time::from_mins;
 use crate::types::Network;
-use crate::utils::from_mins;
 use crate::{DefaultEncoding, Server};
 use crypto::ibe::{generate_seed, public_key_from_master_key};
 use crypto::{ibe, seal_decrypt, seal_encrypt, EncryptionInput, IBEPublicKeys, IBEUserSecretKeys};
