+++ crates/key-server/src/tests/mod.rs
@@ -6,8 +6,8 @@ use crate::key_server_options::{KeyServerOptions, RetryConfig, RpcConfig, Server
 use crate::master_keys::MasterKeys;
 use crate::sui_rpc_client::SuiRpcClient;
 use crate::tests::KeyServerType::Open;
+use crate::time::from_mins;
 use crate::types::Network;
-use crate::utils::from_mins;
 use crate::{DefaultEncoding, Server};
 use crypto::ibe;
 use crypto::ibe::public_key_from_master_key;
