+++ crates/key-server/src/tests/tle.rs
@@ -4,8 +4,9 @@
 use crate::externals::current_epoch_time;
 use crate::tests::externals::sign;
 use crate::tests::SealTestCluster;
+use crate::time::current_epoch_time;
 use crate::valid_ptb::ValidPtb;
-use crate::{current_epoch_time, InternalError};
+use crate::InternalError;
 use crypto::elgamal;
 use fastcrypto::ed25519::Ed25519KeyPair;
 use fastcrypto::traits::KeyPair;
